+++ .gitignore
@@ -1,3 +1,4 @@
 /node_modules
 /dist
 /src/css
+/src/js/main.min.js
