+++ gulpfile.js
@@ -1,4 +1,4 @@
-const{src, dest, task, watch, series, parallel} = require('gulp');
+const{src, dest, watch, series, parallel} = require('gulp');
 const sass = require('gulp-sass');
 const postcss = require('gulp-postcss');
 const autoprefixer = require('autoprefixer');
@@ -6,6 +6,8 @@ const cssnano = require('cssnano');
 const browserSync = require('browser-sync').create();
 const sourcemaps = require('gulp-sourcemaps');
 const del = require('del');
+const webpackStream = require('webpack-stream');
+const rename = require('gulp-rename'); 
 
 function browsersync(){
 	browserSync.init({
@@ -34,6 +36,15 @@ function buildSass(){
 	  .pipe(browserSync.stream())
 };
 
+function buildJs() {
+  return src('src/js/index.js')
+    .pipe(webpackStream(require('./webpack.config')))
+    .pipe(rename('main.min.js'))
+    .pipe(dest('src/js'))
+    .pipe(dest('dist/js'))
+    .pipe(browserSync.stream());
+}
+
 function html(){
 	return src('src/**/*.html')
 	  .pipe(dest('dist/'))
@@ -41,6 +52,7 @@ function html(){
 };
 
 function serve(){
+	watch(['src/js/**/*.js', '!src/js/**/*.min.js'], buildJs);
 	watch('src/scss/**/*.scss', buildSass);
 	watch('src/**/*.html', html);
 }
@@ -54,9 +66,9 @@ function copy(){
 	}).pipe(dest('dist'));
 }
 
-function clianDist(){
+function cleanDist(){
 	return del('dist/**/*', {force:true});
 }
 
-exports.build = series(clianDist, buildSass, html, copy);
-exports.default = series(buildSass, parallel(browsersync, serve));
+exports.build = series(cleanDist, buildSass, buildJs, html, copy);
+exports.default = series([buildSass, buildJs], parallel(browsersync, serve));
