+++ src/index.html
@@ -45,7 +45,55 @@
 		  <div class="header-middle__hero">
 			<h1 class="hero__title"> Создай свой образ вместе с нами</h1>
 			<p> Идеальный образ за 60 минут</p>
-			<button class="btn btn_large"> Записаться онлайн</button>
+			<button class="btn btn_large" data-fancybox data-src="#requect"> Записаться онлайн</button>
+			
+			<div id="request" class='hidden'>
+				<h2>Online-запись</h2>
+				<form class="order-form" id="order-form">
+					<div class="order-form__inner">
+						<div class="form-control">
+							<label>Ваше имя</label>
+							<input name="name" type="text">
+						</div>
+						<div class="form-control">
+							<label>Номер телефона</label>
+							<input name="phone" type="text"  placeholder="+7 (___) ___-__-__" >
+						</div>
+						<div class="form-control">
+							<label>Выберите матера</label>
+							<select name="masterId" value="Не выбран"></select>
+						</div>
+						<div class="form-control">
+							<label>Выберите услугу</label>
+							<select name="serviceId" value="Не выбрана"></select>
+						</div>
+						<div class="form-control">
+							<label>Дата визита</label>
+							<input name="date" type="date">
+						</div>
+						<button type="submit">Отправить</button>
+					</div>
+					
+					<div class="loader loader--style1" title="0">
+					  <svg version="1.1" id="loader-1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
+					   width="40px" height="40px" viewBox="0 0 40 40" enable-background="new 0 0 40 40" xml:space="preserve">
+					  <path opacity="0.2" fill="#000" d="M20.201,5.169c-8.254,0-14.946,6.692-14.946,14.946c0,8.255,6.692,14.946,14.946,14.946
+						s14.946-6.691,14.946-14.946C35.146,11.861,28.455,5.169,20.201,5.169z M20.201,31.749c-6.425,0-11.634-5.208-11.634-11.634
+						c0-6.425,5.209-11.634,11.634-11.634c6.425,0,11.633,5.209,11.633,11.634C31.834,26.541,26.626,31.749,20.201,31.749z"/>
+					  <path fill="#000" d="M26.013,10.047l1.654-2.866c-2.198-1.272-4.743-2.012-7.466-2.012h0v3.312h0
+						C22.32,8.481,24.301,9.057,26.013,10.047z">
+						<animateTransform attributeType="xml"
+						  attributeName="transform"
+						  type="rotate"
+						  from="0 20 20"
+						  to="360 20 20"
+						  dur="0.5s"
+						  repeatCount="indefinite"/>
+						</path>
+					  </svg>
+					</div>
+				</form>
+			</div> <!-- request -->
 		  </div>
 		
 		  <div class="header-middle__girl">
@@ -454,14 +502,12 @@
 				  </ul>
 				</nav>
 			</div>
-
 	</footer>
   </div>	
+  <script src="js/main.min.js"></script>
   <script src="https://cdn.jsdelivr.net/npm/jquery@3.5.1/dist/jquery.min.js"></script>
   <script src="https://cdn.jsdelivr.net/gh/fancyapps/fancybox@3.5.7/dist/jquery.fancybox.min.js"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.8.1/slick.min.js"></script>
-  <script src="menu.js"></script>
-  <script src="main.js"></script>
+  <!-- <script src="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.8.1/slick.min.js"></script> -->
   </body>
   
 </html>
