+++ src/js/index.js
@@ -1,0 +1,5 @@
+import 'slick-carousel/slick/slick';
+//import '@fancyapps/fancybox/dist/jquery.fancybox';
+
+import './main';
+
