+++ src/scss/footer_slyle.scss
@@ -1,3 +1,7 @@
+.hidden{
+	display:none;
+}
+
 .footer-top{
 	background: #F8F6F7;
 	padding:40px 0;
@@ -95,14 +99,24 @@ input{
 	 
 	 input{
 		 width:auto;
-	 }
-		 
+	 }	 
  }
 
+.loader{
+	display:none;
+}
 
+.order-form__inner,
+.hidden h2{
+	width:300px;
+	padding:10px;
+	margin:auto;
+}
 
-
-
+.form-control{
+	width:100%;
+	padding-bottom:15px;
+}
