+++ src/scss/header_slyle.scss
@@ -14,6 +14,7 @@ html,
 body,
 .wrapper{
 	height:100%;
+	scroll-behavior:smooth;
 }
 
 .wrapper{
