+++ webpack.config.js
@@ -1,0 +1,20 @@
+const isProduction = process.env.NODE_ENV === 'production';
+
+module.exports = {
+  mode: process.env.NODE_ENV,
+  devtool: isProduction ? 'none' : '#inline-source-map',
+  module: {
+    rules: [
+      {
+        test: /\.(js)$/,
+        exclude: /(node_modules)/,
+        loader: 'babel-loader'
+      }
+    ]
+  },
+  externals: {
+    $: 'jquery',
+    jQuery: 'jquery',
+    'window.jQuery': 'jquery'
+  },
+};
